vectorClust does a better job of loading saved polygons
saved polygons files now contain information on features and how they were calculated (vcdb.f). When loading polygons, if a feature is required that does not already exist, vectorClust and vcQuickCluster will automatically try to calculate the missing feature based on the feature information. I changed vc_cmf_PartialMean.m so that it adds the necessary information to vcdb.f. This may not work with other features yet.