Added figure learning to reduce variability for RL paper
Using the indirect pathway (and constraining LMAN neurons to have nonnegative
firing rates) it is possible to learn to suppress LMAN firing and thereby reduce
variability.



Michale asked why is this learning so broad? It is much broader that the LMAN
fluctuations that limited normal CAF learning. I don't have an answer to this.