Normalizing reward and eligibility kernels by their integrals
Previously, I was normalizing by the max of the kernels. That was a bad idea
because the overall change in synaptic weight for one motif is proportional to
the integral of the kernels. Now I normalize by the integral of the kernels so
that changing the kernel width does not affect the effective learning rate.
This new normalization made the effective learning rate much lower though, so I
had to raise the learning rates to compensate.

Also shortened kernels to have standard deviation of 20 ms (down from 50 ms).