modifications for plotting distributions of delta weights (dw)
If variable DEBUG_FLAG == 1, sparsemodel_run.m stores a snapshot of weights at
the end of every motif in variable w_all.

Slowed learning rate to 2.5e-5
Lowered threshold for msn activity to zero (no threshold)
Turned off competition

Increased the time msn activity must be above threshold to trigger competition
to 7 ms because this is the total time that a single hvc burst is nonzero. This
won't have any effect now beause competiton is off, but it will allow for better
learning when competition is on again.