Competition on each synapse is proportional to exp(-a * weight)
If activity of an MSN is over an activity threshold for a certain amount of time
in a single motif, all the weights in that MSN are decreased. They are decreased
by a set amount multiplied by a scale factor. The scale factor is like
exp(-a*weight), so large weights receive less competition. This allows the
strongest synapses to grow fast enough to beat the competition while the other
weaker weights cannot beat the competition and are kept low.

Using this competition, MSNs are sparse (active only a few times per song), but
not sparse. At a single time in the song, there are still many MSNs active. Next
I will add lateral inhibition to make MSNs more sparse.