Added Mindlin's model of vocal production
The goal is to have my reinforcement learning model learn the control parameters
(pitch and tension) to produce a song. This is very different from what the
model does now because this Mindlin model is a nonlinear system!

It will be a challenge to see if my model can learn to control this nonlinear
system.

If it can, I think my model will be much more satisfying. It will produce an
actual sound waveform instead of some arbitrary parameters like "pitch". I will
also have demonstrated that the model can overcome one common weakness of RL
models - the ability to control complex systems.