Mindin model with smoothly varying parameters
My original implementation of the Mindlin model had sudden step changes in the
vocal control parameters (air sac pressure and tension on the syrinx) which
caused large transients in the sound waveform. I tried to fix this problem by
making the parameters vary smoothly with linear ramps. When the tension is
changing there is a very low freq component that is messing everything up. Maybe
this will go away when I add the helmholtz resonator?