labeledspecgram.m no longer makes an error when you press buttons on the keyboard
Pushing some keys (j, k, pageup, pagedown, home, end, or the arrow keys) changes the file that is displayed.
Pusing any other keys used to cause an error because it failed to update the file number that is displayed.
Now, pushing other keys does not change the file displayed and does not throw an error.