electro_gui Added option to show events only from zoom box
The event viewer in electro_gui shows the waveforms of events. Until now,
it showed all of the selected events in the current file. If the file is long,
this can cause too many of waveforms to be plotted.

To address this problem, I added an option to display only the event waveforms
from selected events in the current zoom range. This is off by defuault.

To turn on this option, right click on the Event Viewer axes and in the context
menu select "Events From" > "Zoom box"