Include input validation and eliminate silent bailing in display_specgram_quick
display_specgram_quick.m would accept empty signal vectors, and fail to plot
them. It will now throw an error if it is passed an empty signal, as this is
probably not what the caller intended.

display_specgram_quick would also silently bail if the signal became too zoomed
in to generate a taper window. I have increased its verbosity to display a
warning, and then bail.