Opening a new feature branch to improve acquisitionGui reliability
On faster machines, acquisitionGui can crash when triggering on song. This is
due to the daq toolbox using 'pause' to wait for events to occur. When pause is
called within a MATLAB timer, as happens during song triggering, the entire
MATLAB process is suspended during the pause period. This stops all 'background'
threads from progressing, including looking at data from the NiDaq.

These problems are eliminated by eliminating 'pause' calls. To allow this, I
have changed acquisitionGui to use callbacks when events complete. This in turn
is made possible be rewriting the daq toolbox into a class, which can emit
events.