Avoid falling behind on song triggering
Since song triggering peeks at the buffer every second, and the DAQ updates
four times a second. If the DAQ is updating, then song triggering will quit
without peeking. This can lead to song triggering quitting many times in a row,
because it is in sync with the DAQ buffer. I have added a callback to song
triggering if the DAQ buffer is busy, which should eliminate this problem.