Mulling over what to do about concurrency problems
I've done a fair amount of broken-window mending with acquisitionGui, but there
are still many problems when it comes to concurrency. Stressing the GUI by
running multiple experiments makes a concurrency problem much more likely, and
it is hard to keep track of the state of everything.

I think the solution may be an object-oriented architecture, which in Matlab
is the only way to implement events. In addition, the management of global state
through the aa_semaphores (inappropriately named) package is cumbersome and the
source for many bugs.

I am beginning to draft an object-oriented patch, which moves the gui logic into
two new classes: the AcqGuiController class that controlls the whole GUI, and
the SongTriggeredExperiment class that manages the state of each experiment.