Fix errors for song power calculation
I noticed that the way acquisitionGui calculates power in band and out of band
is not correct. I don't know when this error was introduced, but this commit
should fix it. Mapping the desired frequncy range (in Hz) to the PSD values
produced by spectrogram (by index) was incorrect. I have fixed the algorithm to
do this, and have added more checking to the desired frequency range.

I don't know if these changes will alter the behavior and acquisitionGui, some
tweaking may be required to trigger recordings as desired now that the algorithm
has changed.