Alter defaults for SongTriggeredExperiments, get rid of fileTimes
After looking at the restart timer and thinking about my own use of
acquisitionGui, I am making the defaults produced by the restart timer the
defaults for SongTriggeredExperiments. This requies changing a number of
parameters.

fileTimes appears to be unused in acquisitionGui, so I'm deleting it.

I'm also breaking the file writing and directory creation functions off of
create_experiment