Begin to add buffering code
I had eliminated the buffering period from my private branch, but I realized
that song triggering can peek into the past. To stop this from happening, it is
necessary to allow DaqBuffer to collected data for at least as long as it takes
to fulfill a peek (often one second). This commit begins to restore this
functionality.