Rethink song detection in AcqGuiController
acquisitionGui used to use timers to check for song periodically. This timer was
independent of the periodic updates that the DaqBuffer class would receive from
the daq. Instead of having two sets of independent timers running, causing
syncronization and concurrency problems, the song detection timer could be
driven by the DaqBuffer updates.

This is frankly a more natural approach, as asking for more data from the DAQ
should be tied to whether or not there is new data to ask for. It greatly
simplifies the task of checking for song, as it eliminates the need for a timer,
and asking for data happens in a defined part of the DaqBuffer's update.