Sort out timing errors with restart manager
Timers are apprently not deleted or stopped immediately after the stop call, at
least it's not immediate when stop is called within the timer's callback.

I made a second timer object to handle queuing of the next day's restart timer
to give the timer enough time to finish its business and be deleted.