Convert user error into debug message, not error
If the user attempts to a select a segment that exceeds the number of segments
in a file, Parse_segments will throw an error caused by out-of-bounds indexing.
I have added a simple check to make sure that the current selection would not
cause an indexing error, and emit a `fprint` debug statement if it does, rather
than throw an error.